requirements.txt
python-chess
numpy
Pillow
torch
torchvision